backend/backend.old/openai_service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests
python-dotenv
flask-cors
openai
